Moved card enums to their own files Added color attribute for categories
