+++ src/ThroughTheAges.Bot/Enums/Age.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ThroughTheAges.Bot.Enums
+{
+  public enum Age
+  {
+    A = 0,
+    I = 1,
+    II = 2,
+    III = 3,
+    IV = 4
+  }
+}
