+++ src/ThroughTheAges.Bot/Enums/CardType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ThroughTheAges.Bot.Enums
+{
+  public enum CardType
+  {
+    Civil = 1,
+    Military = 2
+  }
+}
